Implementing transfer use case. Some refactorings.

+++ core/src/main/java/com/paulorobertomartins/cleanarch/core/usecases/requestmodel/InputStockRequest.java
@@ -3,11 +3,13 @@ package com.paulorobertomartins.cleanarch.core.usecases.requestmodel;
 import lombok.Builder;
 import lombok.Value;
 
+import java.math.BigDecimal;
+
 @Builder(toBuilder = true)
 @Value
 public class InputStockRequest {
 
     private final String addressLabel;
     private final String productEan;
-    private final Double quantity;
+    private final BigDecimal quantity;
 }
