+++ core/src/main/java/com/paulorobertomartins/cleanarch/core/usecases/requestmodel/TransferStockRequest.java
@@ -3,6 +3,8 @@ package com.paulorobertomartins.cleanarch.core.usecases.requestmodel;
 import lombok.Builder;
 import lombok.Value;
 
+import java.math.BigDecimal;
+
 @Builder(toBuilder = true)
 @Value
 public class TransferStockRequest {
@@ -10,5 +12,5 @@ public class TransferStockRequest {
     private final String addressFromLabel;
     private final String addressToLabel;
     private final String productEan;
-    private final Double quantity;
+    private final BigDecimal quantity;
 }
