+++ core/src/main/java/com/paulorobertomartins/cleanarch/core/usecases/responsemodel/InputStockResponse.java
@@ -3,6 +3,8 @@ package com.paulorobertomartins.cleanarch.core.usecases.responsemodel;
 import lombok.Builder;
 import lombok.Value;
 
+import java.math.BigDecimal;
+
 @Builder(toBuilder = true)
 @Value
 public class InputStockResponse {
@@ -10,5 +12,5 @@ public class InputStockResponse {
     private final Long stockId;
     private final String addressLabel;
     private final String productEan;
-    private final Double quantity;
+    private final BigDecimal quantity;
 }
